backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
psycopg2-binary
python-dotenv
PyJWT
bcrypt
flask-socketio
python-socketio
simple-websocket
pytest
flask-bcrypt


Pillow==10.0.0
Flask-Migrate

python-escpos
pyserial
